Hosting and joining sessions works on local computer.
You can now host and join a session of the game but on 2 instance of the game running on your own computer.
You can see other session that are in LAN but can't join them, I don't know why.

Display Player's name when joining the game.

Some BP functionnality are adding.