Implementation of interaction between past and future self for InteractiveElement and check of overlapping of objects (other than player) in the scene
In this implementation a character can interact with the past flowerpot and see the future flowerpot having its tree appear. This tree is also able to check if it appears overlapping a lever and if it does it can interact with it (for the moment just for test, it calls its function SetVisibility to false)

Modification to classes :
*InteractiveElement
Since an InteractiveElement can have a future counter part, a boolean bCanExistsPastFuture parameter has been added. When it is set to true, the user can choose an AActor to link it with. For that matter the InteractiveElement class now has a parameter AActor* FuturElement.
(For the moment the set is not mandatory and the type of the object with which it will be linked is not checked but it will have to be done to avoid potential errors)

*FlowerPot
A cylinder static mesh is loaded by reference (ConstructorHelpers::FobjectFinder) in the constructor.
FlowerPot has now a UTree* TreeTrunc parameter and a collision capsule for it (this is used to check if the tree is overlapping a lever when it appears)
FlowerPot has a blueprint accessible property to set the tree visible or not
FlowerPot also owns a Seed* parameter which will be needed in case the player will be able to plant several types of seeds.
Method to grow the tree, take out the seed or swap the seed have been implemented

*Seed
A cylinder static mesh is loaded by reference (ConstructorHelpers::FobjectFinder) in the constructor.

Lever
A lamp static mesh is loaded by reference (ConstructorHelpers::FobjectFinder) in the constructor.

Props
As an InteractiveElement it now possesses a static mesh

Tree
Do not inherit from Aactor anymore but of UstaticMeshComponent in order to be created in the flowerPot constructor via a call to PCIP.CreateDefaultSubobject<UTree>